grids ready? fields not
