+++ src/RingGrids/architecture.jl
@@ -1,0 +1,2 @@
+struct DummyArchitecture end
+const DEFAULT_ARCHITECTURE = DummyArchitecture()
